Remove sql folder and references to it
Sql folder made folder hierarchy more confusing
