+++ examples/Microsoft.Spark.CSharp.Examples/Batch/Basic.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using Microsoft.Spark.Sql;
 using static Microsoft.Spark.Sql.Functions;
 
-namespace Microsoft.Spark.Examples.Sql.Batch
+namespace Microsoft.Spark.Examples.Batch
 {
     /// <summary>
     /// A simple example demonstrating basic Spark SQL features.
