+++ examples/Microsoft.Spark.CSharp.Examples/MachineLearning/SentimentAnalysisStream.cs
@@ -7,7 +7,7 @@ using System;
 //using Microsoft.ML.Data;
 using Microsoft.Spark.Sql;
 
-namespace Microsoft.Spark.Examples.Sql.MachineLearning
+namespace Microsoft.Spark.Examples.MachineLearning
 {
     /// <summary>
     /// Example of using ML.NET + .NET for Apache Spark
